Create 3 functions link JS file
Trying to create my "rock-paper-scissors game", I created 3 functions here. 1 is for getting
a randomly generated answer from the computer of either "rock, paper or scissors". The second
function prompts the user to input their answer, and runs the program logic to determine the
winner. I am currently having trouble getting the 'computerSelection' to be accepted by this
playRound() function. Need to work on this.
Lastly, my 3rd function, although not worked on yet, will keep track of the score and end
the game after 5 rounds.