REvise and reset JS file
After trying to put together the game() function with the for loop that would trigger 5
rounds, my code went off the rails and when I tried to comment out the new function,
the basic logic of the game would no longer work. So, i had to revise and change the code to
get the nasic logic of playRound() to work again. As of this commit, I can play one round
of the game and the logic will work. I TOOK OUT the parameters of the playRound funtion:
playRound(playerSelection, computerSelection) --> playRound()