Ensure logic of for loop in game()
I was finally able to get EACH round to take a new prompt from the user, AND a new random
answer from the computer. So now it displays the results of each individual round.
Biggest part left is putting together the IF statement that goes after the loop in game()