Finish game() and declare winner/loser
I completed the for loop that occurs in the game() function. After each instance of
playRound(), the result is given to the console. One thing I still was unsure on how to
do was to account for instances where the user put in an incorrect answer and have it
NOT COUNT as one of the rounds. Looking to student solutions to learn that.
I also added in an IF statement at the end to trigger the correct ine for who won/lost
the game after 5 rounds