Phase out game() function and add check() and disbaleButton()
After attempting to rework the game function and use a while loop that would stop the rounds after
either computer or player hit 5 points, I realized that playRound() was now tied entirely to the buttons,
making the game() function useless. So i created a check() function and inserted it into playRound()
to deliver a game ending message if 5 points was hit. If either side hits 5 points, a disableButton()
function is now alos triggered to prevent the player from continuing to play a round