Add button and mechanism to restart a new game
Originally I tried to to do this through prompt that would pop up and ask the player if
they wanted to play gaain, but then figured it would be more intuitive to make a
'Play again? button appear after either side won the game. I have made the buttton
only appear after the other buttons become disabled and the game is won or lost.
Now working on adding the event listener which will actually trigger the reset