Add stylesheet and title div
I added in the .title div to give some heading to the game page, and then added a stylesheet
to style the title text. I also have been trying to get the play again button to work
correctly, but have not been able to figure it out completley. I can get it to re-enable
the buttons, but it doesn't seem to reset the logic of the game